+++ RELEASES.md
@@ -1779,3 +1779,13 @@ Based on:
 - [typescript v0.9.0-alpha.13] .
 ### Releases
 - [NPM v0.9.0-alpha.13] https://www.npmjs.com/package/cribl-control-plane/v/0.9.0-alpha.13 - .
+
+## 2026-06-11 05:50:28
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.776.0 (2.900.1) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v0.9.0-alpha.14] .
+### Releases
+- [NPM v0.9.0-alpha.14] https://www.npmjs.com/package/cribl-control-plane/v/0.9.0-alpha.14 - .